LIB: Refactor the effective width calculation
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/general/qmesydaq/+/25985
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>

+++ lib/mpsd8.cpp
@@ -639,7 +639,7 @@ quint8 MPSD8::getChannels() const
 	return 8;
 }
 
-quint8 MPSD8::width() const
+quint16 MPSD8::width() const
 {
 	return m_width;
 }
